codes_project/requirements.txt

pandas==1.5.2
flask==2.0.3
singleton-decorator==1.0.0
PyYAML==6.0
redis==4.4.0
gunicorn==20.1.0